Add timeout option for image info block
Create a new timer in viewer to hide image info blocks on trigger; when
set, image info will be displayed on screen until the timeout expires.
When the timeout expires, the UI will redraw and hide info blocks.

The timeout is disabled by default (info block shown forever) but a user
may override it to a fixed time (in seconds) or to a percentage of a
frame time in slideshow mode (eg if slideshow_time=10 and
image_info_timeout=50%, the info block will be hidden after 5 seconds).