[RFC] Downsample when decoding
When used in very constrained systems (eg raspberry pi's) the memory pressure
from Swayimg can be quite significant. It's very easy to bring down a
platform with 512mb of RAM just by loading a few large pictures and
keeping them in memory. Implementing this change would enable loaders to
reduce image size, so that the actual pixels stored in memory are not
more than those that will be rendered to screen.

In my system, using not-very-large images (12mp) I see a reduction in
memory usage of about 1/3, and faster render speeds.