Gallery mode by default for multiple files
calling swayimg for multiple files or a directory doesn't always mean
that the user wants to view only the first file, instead, the user
would want to view multiple files in a gallery view.

inversing the immediate mode switch to gallery for those who want to
view each individual file is better because they would know they chose
the correct files as well.

Signed-off-by: sewn <sewn@disroot.org>